[REFACTOR] Move the challenge unlock flow into a new component
